backend/requirements.txt
# Core Framework
Django==4.2.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter>=24.3
django-environ>=0.11.2
python-decouple==3.8
PyJWT>=2.9.0
bcrypt>=4.2.0
djangorestframework-simplejwt>=5.3.1
drf-spectacular>=0.27.0

# MongoDB Integration
mongoengine>=0.28.2
pymongo>=4.9.0

# Additional utilities
python-dotenv==1.0.1
requests>=2.32.0
celery>=5.4.0
redis>=5.0.8
channels>=4.1.0
channels-redis>=4.2.0
Pillow>=11.0.0
openpyxl>=3.1.5
pandas>=2.2.2
reportlab>=4.2.5

# Development
black>=24.8.0
flake8>=7.1.0
pytest>=8.3.0
pytest-django>=4.9.0